+++ src/parser/enrichers/class/artificer/_module.mjs
@@ -1,6 +1,7 @@
 export { default as ArcaneArmorCreateArmor } from "./ArcaneArmorCreateArmor.mjs";
 export { default as ArcanePropulsionArmorGauntlet } from "./ArcanePropulsionArmorGauntlet.mjs";
 export { default as ArmorModel } from "./ArmorModel.mjs";
+export { default as EldritchCannon } from "./EldritchCannon.mjs";
 export { default as EldritchCannonFlamethrower } from "./EldritchCannonFlamethrower.mjs";
 export { default as EldritchCannonForceBallista } from "./EldritchCannonForceBallista.mjs";
 export { default as EldritchCannonProtector } from "./EldritchCannonProtector.mjs";