+++ src/parser/enrichers/monster/EldritchCannon/_module.mjs
@@ -1,0 +1,3 @@
+export { default as Protector } from "./Protector.mjs";
+export { default as Flamethrower } from "./Flamethrower.mjs";
+export { default as ForceBallista } from "./ForceBallista.mjs";