+++ src/parser/spells/SpellListFactory.mjs
@@ -271,10 +271,8 @@ export default class SpellListFactory {
       },
     };
 
-    if (foundry.utils.isNewerVersion("13", game.version)) {
-      logger.debug(`Updating Journal Page`, { update, page, spells, newSpells });
-      await journal.updateEmbeddedDocuments("JournalEntryPage", [update]);
-    }
+    logger.debug(`Updating Journal Page`, { update, page, spells, newSpells });
+    await journal.updateEmbeddedDocuments("JournalEntryPage", [update]);
 
   }